# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IrwTrafficAI - Pemantau Kemacetan Kota Medan
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TrafficOverviewOverallLevel = typeof TrafficOverviewOverallLevel[keyof typeof TrafficOverviewOverallLevel];


export const TrafficOverviewOverallLevel = {
  lancar: 'lancar',
  sedang: 'sedang',
  padat: 'padat',
  macet_total: 'macet_total',
} as const;

export interface TrafficOverview {
  overallLevel: TrafficOverviewOverallLevel;
  totalRoads: number;
  lancarCount: number;
  sedangCount: number;
  padatCount: number;
  macetCount: number;
  lastUpdated: string;
  summary: string;
}

export type TrafficSegmentLevel = typeof TrafficSegmentLevel[keyof typeof TrafficSegmentLevel];


export const TrafficSegmentLevel = {
  lancar: 'lancar',
  sedang: 'sedang',
  padat: 'padat',
  macet_total: 'macet_total',
} as const;

export interface TrafficSegment {
  id: number;
  name: string;
  from: string;
  to: string;
  level: TrafficSegmentLevel;
  speedKmh: number;
  delayMinutes: number;
  lat: number;
  lng: number;
  lastUpdated: string;
}

export type TrafficHotspotSeverity = typeof TrafficHotspotSeverity[keyof typeof TrafficHotspotSeverity];


export const TrafficHotspotSeverity = {
  sedang: 'sedang',
  parah: 'parah',
  sangat_parah: 'sangat_parah',
} as const;

export interface TrafficHotspot {
  id: number;
  location: string;
  description: string;
  severity: TrafficHotspotSeverity;
  lat: number;
  lng: number;
  estimatedClearTime: string;
  reportedAt: string;
}

export interface TrafficStats {
  totalSegments: number;
  percentageLancar: number;
  percentageSedang: number;
  percentagePadat: number;
  percentageMacet: number;
  worstArea: string;
  bestRoute: string;
  peakHour: string;
  avgSpeedKmh: number;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface ConversationInput {
  title: string;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: Message[];
}

export interface MessageInput {
  content: string;
}

export interface ApiError {
  error: string;
}