# lib/api-zod/src/generated/types/conversationWithMessages.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IrwTrafficAI - Pemantau Kemacetan Kota Medan
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from './message';

export interface ConversationWithMessages {
  id: number;
  title: string;
  createdAt: Date;
  messages: Message[];
}