# lib/api-zod/src/generated/types/trafficHotspot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IrwTrafficAI - Pemantau Kemacetan Kota Medan
 * OpenAPI spec version: 0.1.0
 */
import type { TrafficHotspotSeverity } from './trafficHotspotSeverity';

export interface TrafficHotspot {
  id: number;
  location: string;
  description: string;
  severity: TrafficHotspotSeverity;
  lat: number;
  lng: number;
  estimatedClearTime: string;
  reportedAt: Date;
}