# lib/api-zod/src/generated/types/trafficOverview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IrwTrafficAI - Pemantau Kemacetan Kota Medan
 * OpenAPI spec version: 0.1.0
 */
import type { TrafficOverviewOverallLevel } from './trafficOverviewOverallLevel';

export interface TrafficOverview {
  overallLevel: TrafficOverviewOverallLevel;
  totalRoads: number;
  lancarCount: number;
  sedangCount: number;
  padatCount: number;
  macetCount: number;
  lastUpdated: Date;
  summary: string;
}