# lib/api-zod/src/generated/types/trafficSegment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IrwTrafficAI - Pemantau Kemacetan Kota Medan
 * OpenAPI spec version: 0.1.0
 */
import type { TrafficSegmentLevel } from './trafficSegmentLevel';

export interface TrafficSegment {
  id: number;
  name: string;
  from: string;
  to: string;
  level: TrafficSegmentLevel;
  speedKmh: number;
  delayMinutes: number;
  lat: number;
  lng: number;
  lastUpdated: Date;
}