# lib/api-zod/src/generated/types/trafficStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IrwTrafficAI - Pemantau Kemacetan Kota Medan
 * OpenAPI spec version: 0.1.0
 */

export interface TrafficStats {
  totalSegments: number;
  percentageLancar: number;
  percentageSedang: number;
  percentagePadat: number;
  percentageMacet: number;
  worstArea: string;
  bestRoute: string;
  peakHour: string;
  avgSpeedKmh: number;
}